refactor: add a chat/ layer for card presentation
チャットカード1枚の実装が utils/・data/messages/・applications/ の3〜4ファイルに
散っていた。文言や見た目を直すときに、どのファイルがそのカードを持つのかを毎回
探すことになる。

カードの組み立てと生成（renderTemplate / ChatMessage.create）を module/chat/ に
集め、1カード1ファイルにした。utils/ に残るのは汎用の道具と、config を翻訳して
素材を返すところまで。線は「utils は文字列と行データまで、chat は HTML と
Document まで」で引いている。

あわせて3つ整理した。

- カードの *State 型をスキーマのある data/messages/ へ移した。形の正が1箇所になる
- 8枚が同一形で持っていた addListeners を ChatCardModel の基底へ寄せた。サブクラス
  が持つのは CARD（根の要素・ログ名・失敗時の言語キー）と ACTIONS の2つになる
- utils/chat.ts の DamageApplied は data/messages の AppliedTarget と同じ形だったので
  後者に寄せた

ChatCardModel.ACTIONS を CardActions<never> で宣言しているのは、ハンドラの this が
反変で、基底の型で宣言するとサブクラスの表が代入不可になるため。

武器カードのボタンハンドラは data/ に残っている（architecture.md 課題5）。
これは次のコミットで applications/ へ移す。
