+++ module/data/messages/kai-attack-card.ts
@@ -1,7 +1,18 @@
-import { attachCardActions, type CardActions } from "../../utils/chat-card";
-import { EmokloreSystemDataModel } from "../system-model";
+import type { CardActions } from "../../utils/chat-card";
+import { ChatCardModel } from "./card-model";
 
 const { BooleanField, DocumentUUIDField, NumberField, StringField } = foundry.data.fields;
+
+/** カードの描画・保存に要る状態。スキーマと同じ形 */
+export type KaiAttackCardState = {
+  attackName: string;
+  actorUuid: string | null;
+  mpCost: number;
+  judgeless: boolean;
+  /** 判定の成功数。judgeless のときは固定成功数 */
+  successCount: number | null;
+  damageTotal: number | null;
+};
 
 const defineKaiAttackCardSchema = () => {
   return {
@@ -24,7 +35,7 @@ const defineKaiAttackCardSchema = () => {
  * `emoklore.ts` の init が `ACTIONS` へ登録する。武器カードが `data/` でオーケストレータ化
  * している既知の課題（architecture.md 課題5）を、怪異カードでは繰り返さない。
  */
-export class KaiAttackCardModel extends EmokloreSystemDataModel {
+export class KaiAttackCardModel extends ChatCardModel {
   declare attackName: string;
   declare actorUuid: string | null;
   declare mpCost: number;
@@ -32,23 +43,18 @@ export class KaiAttackCardModel extends EmokloreSystemDataModel {
   declare successCount: number | null;
   declare damageTotal: number | null;
 
+  static override CARD = {
+    root: ".em-kai-attack-card",
+    label: "怪異の攻撃カード",
+    errorKey: "EMOKLORE.ChatMessage.kaiAttack.ActionFailed",
+  };
+
   /** カードのボタン。`data-action` の値と対応する。ハンドラは applications/ 側から登録する */
-  static ACTIONS: CardActions<KaiAttackCardModel> = {};
+  static override ACTIONS: CardActions<KaiAttackCardModel> = {};
 
   static override defineSchema() {
     return defineKaiAttackCardSchema();
   }
 
   static override LOCALIZATION_PREFIXES = ["EMOKLORE.ChatMessage.kaiAttack"];
-
-  /** ボタンに反応する。`renderChatMessageHTML` から呼ばれる（配線は utils/chat-card.ts） */
-  addListeners(html: HTMLElement): void {
-    attachCardActions(html, {
-      root: ".em-kai-attack-card",
-      model: this,
-      actions: KaiAttackCardModel.ACTIONS,
-      label: "怪異の攻撃カード",
-      errorKey: "EMOKLORE.ChatMessage.kaiAttack.ActionFailed",
-    });
-  }
 }
